+++ src/crypto/rand.rs
@@ -5,6 +5,7 @@ pub fn rand_bytes(len: uint) -> Vec<u8> {
     unsafe {
         let mut out = Vec::with_capacity(len);
 
+        ffi::init();
         let r = ffi::RAND_bytes(out.as_mut_ptr(), len as c_int);
         if r != 1 as c_int { fail!() }
