+++ src/ffi.rs
@@ -1,7 +1,9 @@
 #![allow(non_camel_case_types, non_uppercase_statics, non_snake_case)]
 #![allow(dead_code)]
 use libc::{c_void, c_int, c_char, c_ulong, c_long, c_uint, c_uchar, size_t};
+use std::mem;
 use std::ptr;
+use std::rt::mutex::NativeMutex;
 use sync::one::{Once, ONCE_INIT};
 
 pub use bn::BIGNUM;
@@ -182,13 +184,34 @@ extern {}
 #[link(name="wsock32")]
 extern { }
 
+static mut MUTEXES: *mut Vec<NativeMutex> = 0 as *mut Vec<NativeMutex>;
+
+extern fn locking_function(mode: c_int, n: c_int, _file: *const c_char,
+                               _line: c_int) {
+    unsafe {
+        let mutex = (*MUTEXES).get_mut(n as uint);
+
+        if mode & CRYPTO_LOCK != 0 {
+            mutex.lock_noguard();
+        } else {
+            mutex.unlock_noguard();
+        }
+    }
+}
+
 pub fn init() {
     static mut INIT: Once = ONCE_INIT;
 
     unsafe {
         INIT.doit(|| {
             SSL_library_init();
-            SSL_load_error_strings()
+            SSL_load_error_strings();
+
+            let num_locks = CRYPTO_num_locks();
+            let mutexes = box Vec::from_fn(num_locks as uint, |_| NativeMutex::new());
+            MUTEXES = mem::transmute(mutexes);
+
+            CRYPTO_set_locking_callback(locking_function);
         })
     }
 }
